Remove DndListHandle module and component
